refactor: 체스판 출력 관련 메소드를 View로 분리
