+++ src/main/java/softeer2nd/chess/Game.java
@@ -1,10 +1,6 @@
 package softeer2nd.chess;
 
 import softeer2nd.chess.piece.Piece;
-import softeer2nd.chess.util.PieceUtils;
-import softeer2nd.chess.util.StringUtils;
-
-import static softeer2nd.chess.Board.BOARD_LENGTH;
 
 public class Game {
 
@@ -14,27 +10,8 @@ public class Game {
         this.board = board;
     }
 
-    public void print() {
-        System.out.print(showBoard());
-    }
-
     public void initializeBoard() {
         board.initialize();
-    }
-
-    public String showBoard() {
-        StringBuilder builder = new StringBuilder();
-        for (int i = 0; i < BOARD_LENGTH; i++) {
-            for (int j = 0; j < BOARD_LENGTH; j++) {
-                if (board.getBoard().get(i).getPiece(j).getType().equals(PieceUtils.Type.NO_PIECE)) {
-                    builder.append('.');
-                } else {
-                    builder.append(board.getBoard().get(i).getPiece(j).getRepresentation());
-                }
-            }
-            builder.append(StringUtils.NEWLINE);
-        }
-        return builder.toString();
     }
 
     public void move(String pos, Piece piece) {
